requirements.txt
asgiref==3.4.1
Django==3.2.6
django-rest-auth==0.9.5
django-taggit==1.5.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
iso8601==0.1.16
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
